# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ScanFileFormat = typeof ScanFileFormat[keyof typeof ScanFileFormat];


export const ScanFileFormat = {
  stl: 'stl',
  obj: 'obj',
  ply: 'ply',
} as const;

export type ScanJaw = typeof ScanJaw[keyof typeof ScanJaw];


export const ScanJaw = {
  upper: 'upper',
  lower: 'lower',
  both: 'both',
  unknown: 'unknown',
} as const;

export interface Scan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  fileFormat: ScanFileFormat;
  filePath: string;
  fileSize: number;
  vertexCount: number;
  triangleCount: number;
  jaw: ScanJaw;
  /** @nullable */
  patientId?: string | null;
  /** @nullable */
  scannerModel?: string | null;
  uploadedAt: string;
  updatedAt?: string;
}

export type ScanUploadJaw = typeof ScanUploadJaw[keyof typeof ScanUploadJaw];


export const ScanUploadJaw = {
  upper: 'upper',
  lower: 'lower',
  both: 'both',
  unknown: 'unknown',
} as const;

export interface ScanUpload {
  file: Blob;
  name?: string;
  description?: string;
  jaw?: ScanUploadJaw;
  patientId?: string;
  scannerModel?: string;
}

export type ScanUpdateJaw = typeof ScanUpdateJaw[keyof typeof ScanUpdateJaw];


export const ScanUpdateJaw = {
  upper: 'upper',
  lower: 'lower',
  both: 'both',
  unknown: 'unknown',
} as const;

export interface ScanUpdate {
  name?: string;
  description?: string;
  jaw?: ScanUpdateJaw;
  patientId?: string;
  scannerModel?: string;
}

export type ScanStatsFormatBreakdown = {[key: string]: number};

export type ScanStatsJawBreakdown = {[key: string]: number};

export interface ScanStats {
  totalScans: number;
  totalFileSize: number;
  avgVertexCount: number;
  avgTriangleCount: number;
  formatBreakdown: ScanStatsFormatBreakdown;
  jawBreakdown: ScanStatsJawBreakdown;
}

export type MeasurementType = typeof MeasurementType[keyof typeof MeasurementType];


export const MeasurementType = {
  distance: 'distance',
  angle: 'angle',
  area: 'area',
  perimeter: 'perimeter',
  depth: 'depth',
} as const;

export interface Measurement {
  id: number;
  scanId: number;
  type: MeasurementType;
  label: string;
  value: number;
  unit: string;
  /** @nullable */
  pointA?: string | null;
  /** @nullable */
  pointB?: string | null;
  /** @nullable */
  pointC?: string | null;
  createdAt: string;
}

export type MeasurementInputType = typeof MeasurementInputType[keyof typeof MeasurementInputType];


export const MeasurementInputType = {
  distance: 'distance',
  angle: 'angle',
  area: 'area',
  perimeter: 'perimeter',
  depth: 'depth',
} as const;

export interface MeasurementInput {
  type: MeasurementInputType;
  label: string;
  value: number;
  unit: string;
  pointA?: string;
  pointB?: string;
  pointC?: string;
}

export interface Annotation {
  id: number;
  scanId: number;
  label: string;
  /** @nullable */
  notes?: string | null;
  position: string;
  color: string;
  createdAt: string;
}

export interface AnnotationInput {
  label: string;
  notes?: string;
  position: string;
  color: string;
}

export type ExportOptionsFormat = typeof ExportOptionsFormat[keyof typeof ExportOptionsFormat];


export const ExportOptionsFormat = {
  stl: 'stl',
  obj: 'obj',
  ply: 'ply',
} as const;

export interface ExportOptions {
  format: ExportOptionsFormat;
  binary?: boolean;
  includeAnnotations?: boolean;
}

export interface ErrorResponse {
  error: string;
}