# lib/api-zod/src/generated/types/annotation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Annotation {
  id: number;
  scanId: number;
  label: string;
  /** @nullable */
  notes?: string | null;
  position: string;
  color: string;
  createdAt: Date;
}