# lib/api-zod/src/generated/types/annotationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface AnnotationInput {
  label: string;
  notes?: string;
  position: string;
  color: string;
}