# lib/api-zod/src/generated/types/exportOptions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportOptionsFormat } from './exportOptionsFormat';

export interface ExportOptions {
  format: ExportOptionsFormat;
  binary?: boolean;
  includeAnnotations?: boolean;
}