# lib/api-zod/src/generated/types/measurement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementType } from './measurementType';

export interface Measurement {
  id: number;
  scanId: number;
  type: MeasurementType;
  label: string;
  value: number;
  unit: string;
  /** @nullable */
  pointA?: string | null;
  /** @nullable */
  pointB?: string | null;
  /** @nullable */
  pointC?: string | null;
  createdAt: Date;
}