# lib/api-zod/src/generated/types/measurementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementInputType } from './measurementInputType';

export interface MeasurementInput {
  type: MeasurementInputType;
  label: string;
  value: number;
  unit: string;
  pointA?: string;
  pointB?: string;
  pointC?: string;
}