# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanFileFormat } from './scanFileFormat';
import type { ScanJaw } from './scanJaw';

export interface Scan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  fileFormat: ScanFileFormat;
  filePath: string;
  fileSize: number;
  vertexCount: number;
  triangleCount: number;
  jaw: ScanJaw;
  /** @nullable */
  patientId?: string | null;
  /** @nullable */
  scannerModel?: string | null;
  uploadedAt: Date;
  updatedAt?: Date;
}