# lib/api-zod/src/generated/types/scanStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanStatsFormatBreakdown } from './scanStatsFormatBreakdown';
import type { ScanStatsJawBreakdown } from './scanStatsJawBreakdown';

export interface ScanStats {
  totalScans: number;
  totalFileSize: number;
  avgVertexCount: number;
  avgTriangleCount: number;
  formatBreakdown: ScanStatsFormatBreakdown;
  jawBreakdown: ScanStatsJawBreakdown;
}