# lib/api-zod/src/generated/types/scanUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanUpdateJaw } from './scanUpdateJaw';

export interface ScanUpdate {
  name?: string;
  description?: string;
  jaw?: ScanUpdateJaw;
  patientId?: string;
  scannerModel?: string;
}