# lib/api-zod/src/generated/types/scanUpload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental CAD Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanUploadJaw } from './scanUploadJaw';

export interface ScanUpload {
  file: Blob;
  name?: string;
  description?: string;
  jaw?: ScanUploadJaw;
  patientId?: string;
  scannerModel?: string;
}